Acceso a Biblioteca + Ficha Tutoria

+++ Vista/Libro/lista-libro.php
@@ -32,43 +32,62 @@
                                             <th>Nombre</th>
                                             <th>Autor</th>
                                             <th style="text-align: center">Tipo</th>
+                                            <?php if($_SESSION['rol']==4){ ?>
                                             <th style="text-align: center" >Editar</th>
                                             <th style="text-align: center" >Eliminar</th>
+                                            <?php } ?>
                                             <th style="text-align: center" >Ver</th>
-											
-                                            
                                         </tr>
                                     </thead>
                                     <tbody>
-                                    <?php foreach($this->model->ListarLibro() as $r): ?>
-                                        <tr class="odd gradeX">
+                                    <?php 
+                                    if($_SESSION['rol']==4){
+                                        foreach($this->model->ListarLibro() as $r): ?>
+                                            <tr class="odd gradeX">
 											<?php if($r->libro_cantidad_disponible!=0 or $r->libro_tipo==1){ ?>
+                                                <td><?php echo $r->libro_codigo; ?></td>
+                                                <td><?php echo $r->libro_nombre; ?></td>
+                                                <td><?php echo $r->libro_autor; ?></td>
+											     <?php if($r->libro_tipo == 1){ ?>
+												<td style="text-align: center">Virtual</td>
+											     <?php }else{ ?>
+											    <td style="text-align: center">Fisico</td>
+											     <?php } ?>
+                                            
+                                                <td style="text-align: center" class="center"><a href="?c=libro&a=Crud&libro_id=<?php echo $r->libro_id; ?>"><span class="glyphicon glyphicon-pencil"></a></td>
+                                                <td style="text-align: center" class="center"><a onclick="javascript:return confirm('¿Seguro de eliminar este registro?');" href="?c=libro&a=Eliminar&libro_id=<?php echo $r->libro_id; ?>"><span class="glyphicon glyphicon-trash"></span></a></td>
+                                            
+
+                                                <td style="text-align: center" class="center"><a href="?c=libro&a=Perfil&libro_id=<?php echo $r->libro_id; ?>"><span class="glyphicon glyphicon-eye-open"></span></a></td>
+                                            <?php }else{ ?>
+											     <td bgcolor="red"><?php echo $r->libro_codigo; ?></td>
+                                                <td bgcolor="red"><?php echo $r->libro_nombre; ?></td>
+                                                <td bgcolor="red"><?php echo $r->libro_autor; ?></td>
+                                                <?php if($r->libro_tipo == 1){ ?>
+												<td style="text-align: center" bgcolor="red">Virtual</td>
+											     <?php }else{ ?>
+											    <td style="text-align: center" bgcolor="red">Fisico</td>
+											     <?php } ?>
+                                                <td style="text-align: center" bgcolor="red" class="center"><a href="?c=libro&a=Crud&libro_id=<?php echo $r->libro_id; ?>"><span class="glyphicon glyphicon-pencil"></a></td>
+                                                <td style="text-align: center" bgcolor="red" class="center"><a onclick="javascript:return confirm('¿Seguro de eliminar este registro?');" href="?c=libro&a=Eliminar&libro_id=<?php echo $r->libro_id; ?>"><span class="glyphicon glyphicon-trash"></span></a></td>
+                                                <td style="text-align: center" bgcolor="red" class="center"><a href="?c=libro&a=Perfil&libro_id=<?php echo $r->libro_id; ?>"><span class="glyphicon glyphicon-eye-open"></span></a></td>
+											<?php } ?>
+                                        </tr>                                        
+                                    <?php endforeach; 
+                                    }else{
+                                       
+                                     foreach($this->model->ListarLibroVirtual() as $r): ?>
+                                        <tr class="odd gradeX">
+                                            <?php if($r->libro_cantidad_disponible!=0 or $r->libro_tipo==1){ ?>
                                             <td><?php echo $r->libro_codigo; ?></td>
                                             <td><?php echo $r->libro_nombre; ?></td>
                                             <td><?php echo $r->libro_autor; ?></td>
-											<?php if($r->libro_tipo == 1){ ?>
-												<td style="text-align: center">Virtual</td>
-											<?php }else{ ?>
-											    <td style="text-align: center">Fisico</td>
-											<?php } ?>
-                                            <td style="text-align: center" class="center"><a href="?c=libro&a=Crud&libro_id=<?php echo $r->libro_id; ?>"><span class="glyphicon glyphicon-pencil"></a></td>
-                                            <td style="text-align: center" class="center"><a onclick="javascript:return confirm('¿Seguro de eliminar este registro?');" href="?c=libro&a=Eliminar&libro_id=<?php echo $r->libro_id; ?>"><span class="glyphicon glyphicon-trash"></span></a></td>
+                                            <td style="text-align: center">Virtual</td>
                                             <td style="text-align: center" class="center"><a href="?c=libro&a=Perfil&libro_id=<?php echo $r->libro_id; ?>"><span class="glyphicon glyphicon-eye-open"></span></a></td>
-                                            <?php }else{ ?>
-											<td bgcolor="red"><?php echo $r->libro_codigo; ?></td>
-                                            <td bgcolor="red"><?php echo $r->libro_nombre; ?></td>
-                                            <td bgcolor="red"><?php echo $r->libro_autor; ?></td>
-                                            <?php if($r->libro_tipo == 1){ ?>
-												<td style="text-align: center" bgcolor="red">Virtual</td>
-											<?php }else{ ?>
-											    <td style="text-align: center" bgcolor="red">Fisico</td>
-											<?php } ?>
-                                            <td style="text-align: center" bgcolor="red" class="center"><a href="?c=libro&a=Crud&libro_id=<?php echo $r->libro_id; ?>"><span class="glyphicon glyphicon-pencil"></a></td>
-                                            <td style="text-align: center" bgcolor="red" class="center"><a onclick="javascript:return confirm('¿Seguro de eliminar este registro?');" href="?c=libro&a=Eliminar&libro_id=<?php echo $r->libro_id; ?>"><span class="glyphicon glyphicon-trash"></span></a></td>
-                                            <td style="text-align: center" bgcolor="red" class="center"><a href="?c=libro&a=Perfil&libro_id=<?php echo $r->libro_id; ?>"><span class="glyphicon glyphicon-eye-open"></span></a></td>
-											<?php } ?>
+                                            <?php } ?>
                                         </tr>                                        
-                                    <?php endforeach; ?>    
+                                    <?php endforeach; 
+                                    } ?>
                                     </tbody>
                                 </table>
                             </div>
